+++ history.md
@@ -119,4 +119,5 @@
   - *breaking:* Service entries must have owner
   - *breaking:* dropped `IServiceContainerExtensions.Lazy()` support
   - *added:* [AOP](https://en.wikipedia.org/wiki/Aspect-oriented_programming ) support ([details](https://github.com/Sholtee/injector#aspects ))
+  - *added:* Provider recipe
   - *upgraded:* `ProxyGen.NET` to version 2.0.0